requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
bcrypt==4.0.1
python-jose==3.3.0
requests==2.31.0
